pyproject.toml
[tool.poetry]
name = "tesc-public"
version = "0.1.0"
description = "TESC validation: benchmarks and uncertainty/dynamics evaluation"
authors = ["AMAWTA Research <labs@amawtalabs.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
google-genai = "*"
numpy = "*"
matplotlib = "*"
sentence-transformers = "*"
requests = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
